+++ CHANGELOG.md
@@ -79,7 +79,9 @@ When watching fields `[up-watch]`, `[up-autosubmit]`, `[up-switch]` or `[up-vali
 
 ### Validations with `[up-validate]`
 
-- You can now disable validation batching with `up.form.config.batchValidate = false`.
+- You can now disable validation batching globally with `up.form.config.batchValidate = false`, or for individual forms or fields with `up.validate({ batch: false })` or `[up-validate-batch="false"]`.
+- Fields or forms can add additional params to the validation request using the `[up-validate-params]` attribute.
+- Fields or forms can add additional headers to the validation request using the `[up-validate-headers]` attribute.
 - You can now send validation requests to a different server route (issue #486). You can use the new `[up-validate-url]` and `[up-validate-method]` attributes on indivudal fields or on entire forms. Unpoly still guarantees eventual consistency in a form with many concurrent validations.
 - Validation targets can now refer to the changed field with `:origin`. This was possible before, but was never documented.